+++ indico/modules/events/registration/client/js/form/fields/registry.js
@@ -158,7 +158,6 @@ const fieldRegistry = {
     settingsFormDecorators: [choiceFieldsSettingsFormDecorator, singleChoiceSettingsFormDecorator],
     settingsFormInitialData: singleChoiceSettingsInitialData,
     icon: 'dropmenu',
-    renderAsFieldset: true,
   },
   multi_choice: {
     title: Translate.string('Multiple Choice'),